refactor(web): group cms components (part 2) - #95
