+++ apps/web/src/routes/+layout.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
 	import { navigating } from '$app/stores';
-	import { Footer } from '$lib/components/Footer';
-	import { Header } from '$lib/components/Header';
-	import { LoadingOverlay } from '$lib/components/LoadingOverlay';
+	import { Footer } from '$lib/components/cms/Footer';
+	import { Header } from '$lib/components/cms/Header';
+	import { LoadingOverlay } from '$lib/components/cms/LoadingOverlay';
 	import { domStore } from '$lib/stores/dom';
 	import { loadingStore } from '$lib/stores/loading';
 	import { isEmpty } from '@jinen/helpers';
